+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes
 
+## May 20, 2025
+- Added support for resolving NWB file URLs from dandiset path in NwbPage
+
+
 ## May 12, 2025
 - Improved error message formatting for large datasets to display sizes in human-readable form (Issue #187)
 - Added permission error notification for embargoed DANDI datasets (Issue #297)